+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 * Fix some cases where `#{...}--` would fail to parse in selectors.
 
+* Allow a single number to be passed to `saturate()` for use in filter contexts.
+
 * Fix a bug where `**/` would fail to close a loud comment.
 
 ## 1.0.0-alpha.4