+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 * Fix a bug with `@extend` superselector calculations.
 
+* Fix some cases where `#{...}--` would fail to parse in selectors.
+
 * Allow a single number to be passed to `saturate()` for use in filter contexts.
 
 * Fix a bug where `**/` would fail to close a loud comment.